+++ pylon_ros2_camera_component/include/internal/impl/pylon_ros2_camera_gige.hpp
@@ -1611,13 +1611,17 @@ float PylonROS2GigECamera::getTemperature()
 {
     try
     {
-        if ( GenApi::IsAvailable(cam_->TemperatureAbs) )
+        if (GenApi::IsAvailable(cam_->TemperatureAbs))
         {  
-            return static_cast<float>(cam_->TemperatureAbs.GetValue());   
+            return static_cast<float>(cam_->TemperatureAbs.GetValue());
+        }
+        else if (GenApi::IsAvailable(cam_->DeviceTemperature))
+        {
+            return static_cast<float>(cam_->DeviceTemperature.GetValue());
         }
         else 
         {
-             return 0.0;
+            return 0.0;
         }
     }
     catch ( const GenICam::GenericException &e )